@@ -10,12 +10,12 @@
     ],
     "minimum-stability": "stable",
     "require": {
-        "php": ">=7.2"
+        "php": "^7.4"
     },
     "require-dev": {
         "phpmd/phpmd": "^2.7",
         "phpstan/phpstan": "^0.12",
-        "phpunit/phpunit": "^8.4",
+        "phpunit/phpunit": "^9.1",
         "squizlabs/php_codesniffer": "^3.5"
     },
     "autoload": {
